app/clients/opensearch_client.py
# app/clients/opensearch_client.py

from opensearchpy import OpenSearch
import os

opensearch_client = OpenSearch(
    hosts=[{"host": os.getenv("OPENSEARCH_HOST", "localhost"), "port": int(os.getenv("OPENSEARCH_PORT", 9200))}],
    http_auth=(os.getenv("OPENSEARCH_USER", "admin"), os.getenv("OPENSEARCH_PASS", "admin")),
    use_ssl=False,
    verify_certs=False
)
