pyproject.toml
[tool.poetry]
name = "ai-feedback-api"
version = "0.1.0"
description = "KFastAPI + AI 시스템"
packages = [{ include = "app" }]
authors = ["Honey"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12.3"

aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.18"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
asgiref = "3.8.1"
attrs = "25.3.0"
backoff = "2.2.1"
bcrypt = "4.3.0"
build = "1.2.2.post1"
cachetools = "5.5.2"
certifi = "2025.4.26"
charset-normalizer = "3.4.2"
chromadb = "1.0.10"
click = "8.1.8"
colorama = "0.4.6"
coloredlogs = "15.0.1"
dataclasses-json = "0.6.7"
decorator = "5.2.1"
Deprecated = "1.2.18"
dict = "2018.11.19"
distro = "1.9.0"
dnspython = "2.7.0"
durationpy = "0.9"
fastapi = "0.115.9"
filelock = "3.18.0"
flatbuffers = "25.2.10"
frozenlist = "1.6.0"
fsspec = "2025.3.2"
gitdb = "4.0.12"
GitPython = "3.1.44"
google-auth = "2.40.1"
googleapis-common-protos = "1.70.0"
greenlet = "3.2.1"
gritql = "0.2.0"
grpcio = "1.71.0"
h11 = "0.16.0"
h2 = "4.2.0"
hpack = "4.1.0"
httpcore = "1.0.9"
httptools = "0.6.4"
httpx = "0.28.1"
httpx-sse = "0.4.0"
huggingface-hub = "0.30.2"
humanfriendly = "10.0"
hypercorn = "0.17.3"
hyperframe = "6.1.0"
idna = "3.10"
importlib_metadata = "8.6.1"
importlib_resources = "6.5.2"
jiter = "0.9.0"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2025.4.1"
kubernetes = "32.0.1"
langchain = "0.3.25"
langchain-cli = "0.0.36"
langchain-community = "0.3.23"
langchain-core = "0.3.60"
langchain-openai = "0.3.17"
langchain-text-splitters = "0.3.8"
langserve = "0.3.1"
langsmith = "0.3.42"
markdown-it-py = "3.0.0"
marshmallow = "3.26.1"
mdurl = "0.1.2"
mmh3 = "5.1.0"
mpmath = "1.3.0"
multidict = "6.4.3"
mypy_extensions = "1.1.0"
numpy = "2.2.5"
oauthlib = "3.2.2"
onnxruntime = "1.21.1"
openai = "1.79.0"
opentelemetry-api = "1.32.1"
opentelemetry-exporter-otlp-proto-common = "1.32.1"
opentelemetry-exporter-otlp-proto-grpc = "1.32.1"
opentelemetry-instrumentation = "0.53b1"
opentelemetry-instrumentation-asgi = "0.53b1"
opentelemetry-instrumentation-fastapi = "0.53b1"
opentelemetry-proto = "1.32.1"
opentelemetry-sdk = "1.32.1"
opentelemetry-semantic-conventions = "0.53b1"
opentelemetry-util-http = "0.53b1"
orjson = "3.10.18"
overrides = "7.7.0"
packaging = "24.2"
posthog = "4.0.1"
priority = "2.0.0"
propcache = "0.3.1"
protobuf = "5.29.4"
public = "2020.12.3"
pyasn1 = "0.6.1"
pyasn1_modules = "0.4.2"
pydantic = "2.11.4"
pydantic-settings = "2.9.1"
pydantic_core = "2.33.2"
Pygments = "2.19.1"
pymongo = "4.12.1"
PyPika = "0.48.9"
pyproject_hooks = "1.2.0"
pyreadline3 = "3.5.4"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
PyYAML = "6.0.2"
referencing = "0.36.2"
regex = "2024.11.6"
requests = "2.32.4"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
rich = "14.0.0"
rpds-py = "0.24.0"
rsa = "4.9.1"
self = "2020.12.3"
setuptools = "80.3.1"
shellingham = "1.5.4"
six = "1.17.0"
smmap = "5.0.2"
sniffio = "1.3.1"
SQLAlchemy = "2.0.40"
sse-starlette = "1.8.2"
starlette = "0.45.3"
sympy = "1.14.0"
tenacity = "9.1.2"
tiktoken = "0.9.0"
tokenizers = "0.21.1"
tomlkit = "0.13.2"
tqdm = "4.67.1"
typer = "0.15.3"
typing-inspect = "0.9.0"
typing-inspection = "0.4.0"
typing_extensions = "4.13.2"
urllib3 = "2.4.0"
uvicorn = "0.34.2"
watchfiles = "1.0.5"
websocket-client = "1.8.0"
websockets = "15.0.1"
wrapt = "1.17.2"
wsproto = "1.2.0"
yarl = "1.20.0"
zipp = "3.21.0"
zstandard = "0.23.0"
langchain-chroma = "0.2.4"
kafka-python-ng = "^2.2.3"
aiokafka = "^0.10.0"
logging = "^0.4.9.6"
pandas = "^2.2.3"
prometheus-fastapi-instrumentator = "^7.1.0"
celery = "^5.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
